Add files via upload
Criação arquivo auxiliar.
